requirements.txt
# Core Telegram Bot
telethon==1.36.0
cryptg==0.4.0

# Database
asyncpg==0.29.0

# HTTP Client
httpx==0.27.0

# AI Providers
cerebras-cloud-sdk>=1.0.0
openai>=1.12.0

# Utilities
python-dotenv==1.0.0